+++ R/function_ScalePar.R
@@ -1,19 +1,19 @@
 #' Scale 'par.txt' files to different model time step
 #'
-#' `ScalePar` scales time step-dependent parameters in an imported 
+#' `ScalePar` scales time step-dependent parameters and recession coefficients in an imported 
 #' [HYPE 'par.txt'](http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:par.txt) parameter file to a 
 #' new target time step. See HYPE wiki tutorial on [sub-daily time steps](http://www.smhi.net/hype/wiki/doku.php?id=start:hype_tutorials:subdaily_timesteps).
 #' 
 #' @param x List containing HYPE parameters. Typically imported with [ReadPar()].
-#' @param timestep.ratio Numeric, time step scaling factor. Defaults to (1/24) to scale from daily to hourly time steps. To scale from hourly to daily time steps, then use 24.
+#' @param timestep.ratio Numeric, time step scaling factor. Defaults to (1/24) to scale from daily to hourly time steps. To scale from hourly to daily time steps use 24.
 #' @param digits Integer, number of significant digits in scaled parameter values to export. See [signif()].
 #' @param verbose Logical, if \code{TRUE}, then information will be printed.
-#' @param print.par Logical, print known time-scale dependent parameters instead of scaling a parameter list.
+#' @param print.par Logical, print known time-scale dependent parameters and recession coefficients instead of scaling a parameter list.
 #' 
 #' @details
-#' `ScalePar` simply applies a user-chosen scaling factor `timestep.ratio` to all time scale-dependent parameters
-#' in a HYPE parameter list. Parameters are matched against an inbuilt set of parameter names. To see this parameters, call \code{ScalePar(print.par = TRUE)}. 
-#' *[Please notify us](https://github.com/rcapell/HYPEtools/issues) if you find parameters missing.*
+#' `ScalePar` applies a user-specified scaling factor, `timestep.ratio`, to all time scale-dependent parameters and recession coefficients
+#' in a HYPE parameter list. Parameters are matched against an inbuilt set of parameter names. To see these parameters, call \code{ScalePar(print.par = TRUE)}. 
+#' *[Please notify us](https://github.com/rcapell/HYPEtools/issues) if you find any missing parameters.*
 #'
 #' If parameters are not timestep-dependent recession coefficients, then scaling is performed using the ratio between the two time step lengths (e.g. 1/24 when scaling from daily to hourly time steps).
 #' If parameters are timestep-dependent recession coefficients, then scaling is performed using the relationship described in: 
@@ -22,13 +22,16 @@
 #' \code{new_parameter_value = 1 - (1 - old_parameter_value)^time_step_ratio}
 #' 
 #' @return
-#' A [list()] object as supplied in `x`, with parameters re-scaled parameters, or nothing if `print.par = TRUE`.
+#' A [list()] object as supplied in `x`, with re-scaled parameters and recession coefficients, or nothing if `print.par = TRUE`.
+#' 
+#' @seealso
+#' \code{\link{ScaleAquiferData}}
 #' 
 #' @examples
 #' # Import daily HYPE parameter file
-#' hpar <- ReadPar(filename = system.file("demo_model", "par.txt", package = "HYPEtools"))
+#' par <- ReadPar(filename = system.file("demo_model", "par.txt", package = "HYPEtools"))
 #' # Scale to hourly time steps
-#' ScalePar(x = hpar)
+#' ScalePar(x = par)
 #' # Print all time scale-dependent parameters known to the function
 #' ScalePar(print.par = TRUE)
 #'