+++ NEWS.md
@@ -5,6 +5,8 @@
 
 #### Highlights
 * Update ScalePar() function to properly handle recession coefficients
+* Add ScaleAquiferData() function to scale RETRATE recession coefficient
+* Add example AquiferData.txt file to demo model; taken from HYPE wiki: http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:aquiferdata.txt
 
 ## HYPEtools 1.6.1 (2924-01-12)
 *Enhancements and bug fixes*